added dynamic rendering to header component
