+++ src/components/Pages/DashBoard.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import Header from '../Dashboard/Header';
+import Header from './Header';
 import Reports from '../Dashboard/Report';
 import CardList from '../Dashboard/CardList';
 
 const Dashboard = () => (
   <section className="pb-5">
-    <Header />
+    <Header type={false} />
     <Reports />
     <CardList />
   </section>
