+++ src/components/Pages/Details.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Header from './Header';
+
+const Details = () => (
+  <section className="pb-5">
+    <Header type />
+  </section>
+);
+
+export default Details;
