+++ src/components/styles/Dashboard.module.css
@@ -11,6 +11,14 @@
   padding: 1rem;
 }
 
+.arrangeRow {
+  float: right;
+  align-items: center;
+  display: flex;
+  height: 100%;
+  justify-items: space-between;
+}
+
 .row {
   background: var(--primary-color-2);
   margin-top: 1rem;
@@ -85,7 +93,7 @@
   }
 
   .marginCreate {
-   min-width: 0rem;
+    min-width: 0rem;
   }
 
   .staticChart {
